楽曲種別削除 API を実装
