--- src/server/app/Http/Requests/Web/SongType/DeleteRequest.php
@@ -1,24 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace App\Http\Requests\Web\SongType;
-
-use Illuminate\Foundation\Http\FormRequest;
-
-class DeleteRequest extends FormRequest
-{
-    /**
-     * Get the validation rules that apply to the request.
-     *
-     * @return array<string, array<string>>
-     */
-    public function rules(): array
-    {
-        return [
-            'song_type_id' => [
-                'required',
-            ],
-        ];
-    }
-}
