+++ src/server/tests/Feature/Api/SongType/DeleteSongTypeTest.php
@@ -1,0 +1,47 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Tests\Feature\Api\SongType;
+
+use Mockery;
+use Mockery\MockInterface;
+use PHPUnit\Framework\Attributes\Test;
+use SongType\Domain\Models\SongTypeId;
+use SongType\Domain\Models\SongTypeRepositoryInterface;
+use SongType\Route\SongTypeRouteMap;
+use Tests\TestCase;
+
+class DeleteSongTypeTest extends TestCase
+{
+    // TODO モックやめる
+    private MockInterface&SongTypeRepositoryInterface $repository;
+
+    public function setUp(): void
+    {
+        parent::setUp();
+
+        $this->repository = Mockery::mock(SongTypeRepositoryInterface::class);
+
+        $this->app->bind(SongTypeRepositoryInterface::class, fn (): SongTypeRepositoryInterface => $this->repository);
+    }
+
+    #[Test]
+    public function canDelete(): void
+    {
+        $uuid = $this->generateUuid();
+
+        $this->repository->shouldReceive('delete')
+            ->with(Mockery::on(fn (SongTypeId $arg) => $arg->value === $uuid))
+            ->once();
+
+        $this->delete(route(SongTypeRouteMap::Delete, $uuid))
+            ->assertStatus(204);
+    }
+
+    #[Test]
+    public function emptyParameters(): void
+    {
+        $this->markTestSkipped('TODO 実装する');
+    }
+}
